+++ bindings/python/examples/secret_manager/ledger_nano.py
@@ -5,7 +5,7 @@ import os
 load_dotenv()
 
 # In this example we will get the ledger status and generate an address
-# To use the ledger nano simulator clone https://github.com/iotaledger/ledger-shimmer-app, run `git submodule init && git submodule update --recursive`,
+# To use the ledger nano simulator clone https://github.com/iotaledger/ledger-iota-app, run `git submodule init && git submodule update --recursive`,
 # then `./build.sh -m nanos|nanox|nanosplus -s` and use `True` in
 # `LedgerNanoSecretManager(True)`.